Added another B&H implementation in MATLAB
This version is contained in a single .m file.  There are other files in
this directory, but I'm not sure if they are important.

Also added a Python driver script that constructs a simple MATLAB script to
call the Mie code and print the results.  The Python code then parses
the MATLAB output and returns the result as Python complex numbers.